consolidate unnecessary components
