+++ src/components/HowToPlay/HowToPlayDialog.module.css
@@ -5,3 +5,7 @@
 .dialog .cancel.push:global(.mdc-dialog__button) {
   margin-right: auto;
 }
+
+.bigBoard {
+  padding: 1.5em 1.5em 0 0;
+}
