--- src/components/HowToPlay/Step/HowToPlayStep.module.css
@@ -1,3 +1,0 @@
-.bigBoard {
-  padding: 1.5em 1.5em 0 0;
-}
